Split out `recordPurchase` and `recordAddToCart`
