+++ client/lib/analytics/test/cart.js
@@ -2,6 +2,7 @@
  * Internal dependencies
  */
 import analytics from 'lib/analytics';
+import { recordAddToCart } from 'lib/analytics/record-add-to-cart';
 import { getAllCartItems } from 'lib/cart-values/cart-items';
 import { recordEvents } from 'lib/analytics/cart';
 
@@ -11,8 +12,11 @@ jest.mock( 'lib/analytics', () => ( {
 		tracks: {
 			recordEvent: jest.fn(),
 		},
-		recordAddToCart: jest.fn(),
 	},
+} ) );
+jest.mock( 'lib/analytics/record-add-to-cart', () => ( {
+	__esModule: true,
+	recordAddToCart: jest.fn(),
 } ) );
 jest.mock( 'lib/cart-values/cart-items', () => ( { getAllCartItems: jest.fn() } ) );
 
@@ -58,7 +62,7 @@ describe( 'recordEvents', () => {
 		recordEvents( previousCart, nextCart );
 
 		expect( analytics.tracks.recordEvent ).not.toHaveBeenCalled();
-		expect( analytics.recordAddToCart ).not.toHaveBeenCalled();
+		expect( recordAddToCart ).not.toHaveBeenCalled();
 	} );
 
 	it( 'records an add event when an item is added', () => {
@@ -71,8 +75,8 @@ describe( 'recordEvents', () => {
 			'calypso_cart_product_add',
 			domainRegNoExtra
 		);
-		expect( analytics.recordAddToCart ).toHaveBeenCalledTimes( 1 );
-		expect( analytics.recordAddToCart ).toHaveBeenCalledWith( { cartItem: domainReg } );
+		expect( recordAddToCart ).toHaveBeenCalledTimes( 1 );
+		expect( recordAddToCart ).toHaveBeenCalledWith( { cartItem: domainReg } );
 	} );
 
 	it( 'records a remove event when an item is removed', () => {
@@ -85,7 +89,7 @@ describe( 'recordEvents', () => {
 			'calypso_cart_product_remove',
 			domainRegNoExtra
 		);
-		expect( analytics.recordAddToCart ).not.toHaveBeenCalled();
+		expect( recordAddToCart ).not.toHaveBeenCalled();
 	} );
 
 	it( 'records an add and a remove event when items are added and removed', () => {
@@ -102,8 +106,8 @@ describe( 'recordEvents', () => {
 			'calypso_cart_product_add',
 			privateRegNoExtra
 		);
-		expect( analytics.recordAddToCart ).toHaveBeenCalledTimes( 1 );
-		expect( analytics.recordAddToCart ).toHaveBeenCalledWith( { cartItem: privateReg } );
+		expect( recordAddToCart ).toHaveBeenCalledTimes( 1 );
+		expect( recordAddToCart ).toHaveBeenCalledWith( { cartItem: privateReg } );
 	} );
 } );
